requirements.txt
# Google Drive & Sheets Integration
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Data Processing & Analysis
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2

# Snowflake Connection
snowflake-connector-python==3.7.0

# Reporting & Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
